examples/requirements.txt
matplotlib
seaborn
torcheval
torchvision
tqdm
jupyterlab
git+https://github.com/eml-eda/plinio.git@main#egg=plinio

